+++ src/serviceDefinition/verify.ts
@@ -1,9 +1,27 @@
-import { red, bold } from 'colors/safe';
+import { red, bold, green, blue } from 'colors/safe';
 
-import { ResilienceScoreServiceState } from './types';
+import { Task, TaskState, ResilienceScoreServiceState } from './types';
 import { abortExecutionWithError } from '../errors';
 import { loadServiceDefinition } from './loading';
 import { executeApiCall } from '../api';
+
+const taskSuffix: Record<TaskState, string> = {
+  'SUCCESS': 'ok',
+  'FAILURE': 'failure',
+  'PENDING': 'pending'
+};
+
+const taskColor: Record<TaskState, typeof red> = {
+  'SUCCESS': green,
+  'FAILURE': red,
+  'PENDING': blue
+};
+
+const taskOrder: Record<TaskState, number> = {
+  'SUCCESS': 0,
+  'PENDING': 1,
+  'FAILURE': 2
+};
 
 export async function verify(serviceDefinitionPath: string) {
   const serviceDefinition = await loadServiceDefinition(serviceDefinitionPath);
@@ -14,9 +32,12 @@ export async function verify(serviceDefinitionPath: string) {
       path: `/api/resilience-score/service/${serviceDefinition.id}`,
     });
     const state = (await response.json()) as ResilienceScoreServiceState;
-    printState(state);
 
-    if (state.desiredResilienceLevel !== state.actualResilienceLevel) {
+    printResilienceLevel(state);
+    console.log();
+    printTaskList(state);
+
+    if (hasAchievedDesiredLevel(state)) {
       process.exit(1);
     }
   } catch (e) {
@@ -24,11 +45,9 @@ export async function verify(serviceDefinitionPath: string) {
   }
 }
 
-function printState(state: ResilienceScoreServiceState) {
-  const desiredLevelAchieved = state.desiredResilienceLevel === state.actualResilienceLevel;
-
+function printResilienceLevel(state: ResilienceScoreServiceState) {
   let color: typeof red = s => s;
-  if (!desiredLevelAchieved) {
+  if (!hasAchievedDesiredLevel(state)) {
     color = red;
   }
 
@@ -36,4 +55,33 @@ function printState(state: ResilienceScoreServiceState) {
   console.log(color(bold('================')));
   console.log('Desired: %s', state.desiredResilienceLevel);
   console.log(color('Actual:  %s'), state.actualResilienceLevel);
+}
+
+function printTaskList(state: ResilienceScoreServiceState) {
+  console.log(bold(`Tasks`));
+  console.log(bold('====='));
+
+  const sortedTasks = state.tasks
+    .slice()
+    .sort((a, b) => taskOrder[a.state] - taskOrder[b.state]);
+
+  const countByType: Record<TaskState, number> = {
+    'SUCCESS': 0,
+    'FAILURE': 0,
+    'PENDING': 0
+  };
+
+  for (const task of sortedTasks) {
+    countByType[task.state]++;
+    console.log(taskColor[task.state](` - %s (%s)`), task.name, taskSuffix[task.state]);
+  }
+
+  console.log();
+  console.log(taskColor.SUCCESS(`Ok:      %d`), countByType.SUCCESS);
+  console.log(taskColor.PENDING(`Pending: %d`), countByType.PENDING);
+  console.log(taskColor.FAILURE(`Failure: %d`), countByType.FAILURE);
+}
+
+function hasAchievedDesiredLevel(state: ResilienceScoreServiceState) {
+  return state.desiredResilienceLevel === state.actualResilienceLevel;
 }